refactor: allow more frequent updates
Instead of speeding up patterns by increasing the tick rate, we use a
standard tick rate now (100ms) and speed up the patterns by a speed
factor, which each pattern can use individually. And base the state
on "now" instead of counting ticks, which may occur more or less
frequent.

So we derive patterns from the state in time, and can control the speed
by a dedicated value.